openhands-internal-mcp-client/pyproject.toml
[project]
name = "openhands-internal-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.25.0",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.0.0",
]
